feat: delete event OK - trying to get author.email for event

+++ back/src/resolvers/event.resolver.ts
@@ -2,6 +2,7 @@
 /* eslint-disable no-console */
 import Event from '../models/Event';
 import { IEvent, IInputEvent, IId } from '../interface/event.interface';
+import { EventModel } from '../models/Event';
 
 const eventResolver = {
   Query: {
@@ -10,12 +11,12 @@ const eventResolver = {
       parent: any,
       { _id }: { _id: string },
       context: any
-    ): Promise<IEvent | null> => {
+    ): Promise<EventModel | null> => {
       if (!context) {
         return null;
       }
       const event = await Event.findOne({ _id });
-      return <IEvent>event;
+      return event;
     },
     getEvents: async (parent: any, args: any, ctx: any) => {
       console.log('Context', ctx.user);
@@ -33,16 +34,20 @@ const eventResolver = {
       _: unknown,
       args: { input: IInputEvent },
       ctx: any
-    ): Promise<IEvent |void> => {
+    ): Promise<EventModel |void> => {
       if (ctx && ctx.user){
-        const event = await Event.create(args.input);
-        return <IEvent>event;
+        const { firstname, lastname, email } = ctx.user;
+        const event = await Event.create({...args.input, author: {firstname, lastname, email}});
+        return event;
       }
       
     },
-
-    deleteEvent: async (_: unknown, args: { input: IId }) => {
-      await Event.deleteOne({ _id: args.input });
+    deleteEvent: async (_: unknown, args: { input: IId }, ctx: any) => {
+      const event = await Event.findOne({ _id: args.input });
+      if (!event) throw new Error("This event does not exist");
+      if (ctx.user.email === event.author.email) {
+        await Event.deleteOne({ _id: args.input });
+      }
     },
   },
 };
